machine_learning_final_project
Customer churn is a critical metric in any customer-centric business. It refers to the percentage of customers who stop using a company's product or service over a given period. This project aims to predict whether a customer is likely to churn based on historical customer data using K-Nearest Neighbors (KNN) and Naive Bayes machine learning algorithms.